Created workable, separate Controllers - a Controller for html and a RestController for api endpoints.
